+++ datafusion/core/tests/parquet/mod.rs
@@ -49,7 +49,6 @@ mod filter_pushdown;
 mod page_pruning;
 mod row_group_pruning;
 mod schema;
-mod schema_adapter;
 mod schema_coercion;
 
 #[cfg(test)]
